+++ app/src/main/AndroidManifest.xml
@@ -17,18 +17,21 @@
         android:usesCleartextTraffic="true">
         <activity
             android:name=".ui.login.view.IntroActivity"
-            android:exported="false" />
+            android:exported="true"
+            android:theme="@style/AppTheme.Splash"
+            android:screenOrientation="portrait">
+
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+
+        </activity>
         <activity
             android:name=".ui.login.view.LoginActivity"
             android:exported="true"
             android:screenOrientation="portrait"
-            android:theme="@style/AppTheme.Splash"
             tools:ignore="LockedOrientationActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name=".MainActivity"
